+++ scanpipe/tests/pipes/test_scancode.py
@@ -40,6 +40,7 @@ from scanpipe.models import CodebaseResource
 from scanpipe.models import DiscoveredDependency
 from scanpipe.models import DiscoveredPackage
 from scanpipe.models import Project
+from scanpipe.pipes import input
 from scanpipe.pipes import output
 from scanpipe.pipes import scancode
 from scanpipe.pipes.input import copy_input
@@ -426,7 +427,7 @@ class ScanPipeScancodePipesTest(TestCase):
     def test_scanpipe_pipes_scancode_load_inventory_from_toolkit_scan(self):
         project = Project.objects.create(name="Analysis")
         input_location = self.data_location / "asgiref-3.3.0_toolkit_scan.json"
-        scancode.load_inventory_from_toolkit_scan(project, input_location)
+        input.load_inventory_from_toolkit_scan(project, input_location)
         self.assertEqual(18, project.codebaseresources.count())
         self.assertEqual(2, project.discoveredpackages.count())
         self.assertEqual(4, project.discovereddependencies.count())
@@ -435,13 +436,13 @@ class ScanPipeScancodePipesTest(TestCase):
         project = Project.objects.create(name="1")
         input_location = self.data_location / "asgiref-3.3.0_scanpipe_output.json"
         scan_data = json.loads(input_location.read_text())
-        scancode.load_inventory_from_scanpipe(project, scan_data)
+        input.load_inventory_from_scanpipe(project, scan_data)
         self.assertEqual(18, project.codebaseresources.count())
         self.assertEqual(2, project.discoveredpackages.count())
         self.assertEqual(4, project.discovereddependencies.count())
 
         # Load again to ensure there is no duplication
-        scancode.load_inventory_from_scanpipe(project, scan_data)
+        input.load_inventory_from_scanpipe(project, scan_data)
         self.assertEqual(18, project.codebaseresources.count())
         self.assertEqual(2, project.discoveredpackages.count())
         self.assertEqual(4, project.discovereddependencies.count())
@@ -450,7 +451,7 @@ class ScanPipeScancodePipesTest(TestCase):
         project2 = Project.objects.create(name="2")
         output_file = output.to_json(project=project)
         scan_data = json.loads(output_file.read_text())
-        scancode.load_inventory_from_scanpipe(project2, scan_data)
+        input.load_inventory_from_scanpipe(project2, scan_data)
         self.assertEqual(18, project2.codebaseresources.count())
         self.assertEqual(2, project2.discoveredpackages.count())
         self.assertEqual(4, project2.discovereddependencies.count())
@@ -459,14 +460,14 @@ class ScanPipeScancodePipesTest(TestCase):
         project = Project.objects.create(name="1")
         input_location = self.data_location / "flume-ng-node-d2d.json"
         scan_data = json.loads(input_location.read_text())
-        scancode.load_inventory_from_scanpipe(project, scan_data)
+        input.load_inventory_from_scanpipe(project, scan_data)
         self.assertEqual(57, project.codebaseresources.count())
         self.assertEqual(0, project.discoveredpackages.count())
         self.assertEqual(0, project.discovereddependencies.count())
         self.assertEqual(18, project.codebaserelations.count())
 
         # Load again to ensure there is no duplication
-        scancode.load_inventory_from_scanpipe(project, scan_data)
+        input.load_inventory_from_scanpipe(project, scan_data)
         self.assertEqual(57, project.codebaseresources.count())
         self.assertEqual(18, project.codebaserelations.count())
 
@@ -474,7 +475,7 @@ class ScanPipeScancodePipesTest(TestCase):
         project = Project.objects.create(name="Analysis")
         filename = "package_assembly_codebase.json"
         project_scan_location = self.data_location / "scancode" / filename
-        scancode.load_inventory_from_toolkit_scan(project, project_scan_location)
+        input.load_inventory_from_toolkit_scan(project, project_scan_location)
 
         project.discoveredpackages.all().delete()
         self.assertEqual(0, project.discoveredpackages.count())