+++ CHANGELOG.rst
@@ -34,6 +34,9 @@ v33.0.0 (unreleased)
 - Improve the speed of Project ``reset`` and ``delete`` using the _raw_delete model API.
   https://github.com/nexB/scancode.io/issues/729
 
+- Add support for XLSX input in the ``load_inventory`` pipeline.
+  https://github.com/nexB/scancode.io/issues/735
+
 v32.2.0 (2023-04-25)
 --------------------